# 296:
- slightly refactored the TestSelectListProperty class to include a
setUp function which initializes one 'SINGLE SELECT LIST' property model
and property instance
- this drastically simplified the tests
- Pycharm was complaining about using set() and was suggesting it can
be replaced with set literal. A quick google search showed that the set
literal is indeed faster. So now instead of set() we are using {a,b,c}
- added new function called test_set_value_in_options() which tests
whether you can correctly set a value from the list of options on the
property instance. This raises the coverage for this class to 100 % !
- we can now set the value of property instance to None, which would
empty it from any option. Very useful to have!