296 increase tests coverage (#297)
* # 296:

- skipped the tests related to ReferenceProperty as the property type is
now deprecated
- added tests to MultiReferenceProperty, coverage for it is now 100%!
- we are not skipping the test_create_and_delete_service() anymore, as
the bug related to it has been long fixed

* # 296:

- slightly refactored the TestSelectListProperty class to include a
setUp function which initializes one 'SINGLE SELECT LIST' property model
and property instance
- this drastically simplified the tests
- Pycharm was complaining about using set() and was suggesting it can
be replaced with set literal. A quick google search showed that the set
literal is indeed faster. So now instead of set() we are using {a,b,c}
- added new function called test_set_value_in_options() which tests
whether you can correctly set a value from the list of options on the
property instance. This raises the coverage for this class to 100 % !
- we can now set the value of property instance to None, which would
empty it from any option. Very useful to have!

* # 296:

- changed all the tests for services to make the call on the scope
instead of the client such that the functions in scope also get covered.
Coverage is now 100% in scope

* # 296:

- added tests for parts, properties, scope and services in order to
improve all the conditional clauses inside the Client class - coverage
for client.py is not at 99%

* # 296:

- increased the tests coverage for services.py to 89%! Not sure how to
tackle get_notebook_url()

* # 296:

- fixed travis complaints

* - adapted the Changelog

* small updates to the documentation in the client (to include IllegalArumentError) and added temp_chdir() to enable a temporary directory usage during testing.