upgrade-sim-scripts-to-python-3-12 (#1416)
* Upgrade default Python version for service execution

The ServiceEnvironmentVersion has been updated to Python 3.12 as default, and Python 3.9 as legacy. The functionality to work with Python notebooks has been deprecated. Corresponding changes have been made to the tests and client.py files to align with this update.

* Update test_services for readability and structure

Reformatted the importing section and function parameters in test_services.py to increase readability. Also added a line to skip tests when using Travis or Github Actions due to authorization issues. This change improves structure and test run conditional checks.

---------

Co-authored-by: Jochem Berends <jochem.berends@ke-works.com>