support-the-multi-attachmentviewer (#1434)
* Add multi attachment viewer widget support to Pykechain

Updated Pykechain to include support for a new multi attachment viewer widget. This update includes changes to the widgets manager, the enums file, and the widget models file to accommodate the new widget type. Now, users can add and customize multi attachment viewer widgets in Pykechain with the provided methods.

* Update pykechain/models/widgets/widgets_manager.py

* Add multiattachment viewer widget and update dependent versions

In this commit, a multiattachment viewer widget has been added to the codebase. This update is made to align with the KE-chain-2 backend. Additionally, dependent versions for development have been updated.

---------

Co-authored-by: Jochem Berends <jochem.berends@ke-works.com>