1443-strip-progress-elements-from-the-metapanel (#1445)
* Remove unused 'progress' attribute from Widget model

The 'progress' attribute, previously extracted from JSON data, was found to be redundant and was therefore removed for code efficiency

* Mark ProgressWidget as deprecated

The progress widget in KE-chain is marked as obsolete, effective from June 2024. Warnings of deprecation are added in relevant methods such as 'create_widget' and '_validate_widget'. Also, deprecation notes are added to the widget's docstrings.

* Add deprecation warning for progress widget

A deprecation warning has been added to the widget validation function in the pykechain client. This warning notifies users that the progress widget will not be available in KE-chain from June 2024 onwards.

* Update test widget attributes in cassettes

The TestWidgets.test_widget_attributes.json file has been updated. The http_interactions have been adjusted with new requests and responses towards different API endpoints, which include key revisions on the headers and body. This should improve the schema tests accuracy and reliability.

* Remove unnecessary whitespace in client.py

An extra line-break in the 'py

* Remove deprecated progress widget warnings

The warnings related to the progress widget, which is no longer available in KE-chain from June 2024, have been removed. This includes

* Add deprecation warnings for Progress widgets

We've added warnings for developers using Progress widgets, informing them about the deprecation from June 2024 onwards. This change is in preparation for the complete removal of the Progress widgets functionality in future updates.

---------

Co-authored-by: Jochem Berends <jochem.berends@ke-works.com>