Added option to include PDF appendices inline in the activity model
Added a new parameter "include_pdf_appendices_inline" to the download_as_pdf method of the activity model, enabling inline rendering of PDF appendices. Refactored the respective unit tests for this new functionality. If the new parameter is set to False, zip files are returned instead of PDFs, containing all the non-renderable attachments.