Refactor error handling for scope cloning (#1458)
* Refactor error handling for scope cloning

Changed from manual status code checking to using `raise_for_status()` for better exception handling. This improves code readability and leverages built-in exception features of the requests library.

* Fix status code check in error handling

Changed the HTTP status code check from `err.response` to `response` in the error handling block. This ensures that the correct response object is used when raising a `ForbiddenError` for a forbidden scope clone attempt.

* Refine scope cloning logic and error handling

This commit enhances the scope cloning logic by checking for specific response status codes. It ensures that the function returns the appropriate response for both `created` and `accepted` statuses while raising an error for unexpected responses.

* changelog item Accept 202 response for scope clone

Updated the changelog to reflect that a 202 Accepted HTTP response is now also accepted for the scope clone operation. This change is tracked under issue #1459.

---------

Co-authored-by: Kamil Niski <kamil.niski@ke-works.com>
Co-authored-by: jberends <jberends@jbits.nl>
Co-authored-by: Jochem Berends <jochem.berends@ke-works.com>