Add signature property (#1467)
* Add SignatureProperty class to handle stored file signatures.

Introduced a new `SignatureProperty` class, enabling manipulation and retrieval of stored file signatures. Updated the models and enum to incorporate the new property type, ensuring complete integration within the existing framework.

* Add tests for creating and uploading to signature properties

This commit introduces a new test suite `TestPropertiesSignatureProperty`. It includes tests to create signature property and a placeholder for uploading a file to the property, extending the test coverage for signature-related operations.

* Fix value serialization in PropertyReference

Serialize the value to a list and handle first element to fix assignment issues in PropertyReference. Additionally, updated a test with corrected filename and provided more clarity on the skipped test.

* Optimize test properties handling in `tests/test_properties.py`

Refactor import ordering for clarity and add setup/teardown steps in signature property test. Utilize pre-created stored file for signature property testing, enhancing maintainability and reducing code repetition.

* Refactor signature property tests and stored file handling

Removed outdated cassette file and refactored signature property tests into a new module. Updated stored file handling to support in-memory image files for uploads. Minor changes to property reference method to access stored files by ID.

* Unskip and fix test_upload_new_signature_to_property

Remove the unnecessary skip annotation line and ensure the `test_upload_new_signature_to_property` test runs properly. This improves test coverage and helps verify that uploading a new signature to a property works as expected.

* Remove unused import Base from property_reference.py

The import of the 'Base' module was removed as it was not used in the code. This helps to clean up the code and avoid unnecessary dependencies. An extra newline was also added to improve readability.

* Add CHANGELOG

Include details about the new signature property feature which handles signatures better and stores them as StoredFiles. This update precedes the version v4.13.0 change log.

* D102 fix Add docstring to `value` property method

Added a docstring to the `value` property method in the `property_reference.py` file. This enhances code readability by providing a clear explanation of the method's purpose and return type.

* Improve Code Quality W0106 Fix test assertion syntax in test_properties.py

Removed unnecessary parentheses around an assertEqual statement to correct the syntax. This change ensures the test properly verifies the unit value in the property model. Fixing issue-a87a2d778b1fe70b9fff62d818103d09 in codacy

* Remove commented-out 400 error message in test

The commented-out error message in `test_upload_new_signature_to_property` is removed to clean up the code. This improves readability and maintains the focus on functional parts of the test.

---------

Co-authored-by: Jochem Berends <jochem.berends@ke-works.com>